edited final project
